requirements.txt
requests==2.32.3
pymongo==4.11.3
argparse==1.4.0
beautifulsoup4==4.13.3
flask==3.1.0
python-dotenv==1.1.0
alerts_in_ua==0.2.7
numpy==2.2.4
seaborn==0.13.2
matplotlib==3.10.1
pandas==2.2.3
nltk==3.9.1
scikit-learn==1.6.1
flask-cors==5.0.1
uwsgi==2.0.29
pytest==8.3.5
mongomock==4.3.0
imblearn==0.0
python-telegram-bot==22.0
apscheduler==3.11.0